Make route system and refactor the code
