+++ Tigran_Nalbandyan/Node.js/School/db/mysql.js
@@ -1,0 +1,11 @@
+const mysql = require('mysql');
+const connection = mysql.createConnection({
+    host: 'localhost',
+    user: 'root',
+    password: 'root',
+    database: 'School'
+});
+
+connection.connect();
+
+module.exports = connection;
