+++ Tigran_Nalbandyan/Node.js/School/test/test.js
@@ -2,7 +2,7 @@ process.env.NODE_ENV = 'test';
 
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const server = require('../index.js');
+const server = require('../app.js');
 const should = chai.should();
 const db = require('../db.js');
